pull each goal into its own file and combine them in CLI
